+++ packages/backend/src/server/api/endpoint-list.ts
@@ -322,6 +322,7 @@ export * as 'invite/delete' from './endpoints/invite/delete.js';
 export * as 'invite/limit' from './endpoints/invite/limit.js';
 export * as 'invite/list' from './endpoints/invite/list.js';
 export * as 'juice/public-settings' from './endpoints/juice/public-settings.js';
+export * as 'juice/ranking' from './endpoints/juice/ranking.js';
 export * as 'meta' from './endpoints/meta.js';
 export * as 'miauth/gen-token' from './endpoints/miauth/gen-token.js';
 export * as 'mute/create' from './endpoints/mute/create.js';