+++ packages/misskey-js/src/autogen/apiClientJSDoc.ts
@@ -3762,6 +3762,17 @@ declare module '../api.js' {
      * 
      * **Credential required**: *No*
      */
+    request<E extends 'juice/ranking', P extends Endpoints[E]['req']>(
+      endpoint: E,
+      params: P,
+      credential?: string | null,
+    ): Promise<SwitchCaseResponseType<E, P>>;
+
+    /**
+     * No description provided.
+     * 
+     * **Credential required**: *No*
+     */
     request<E extends 'meta', P extends Endpoints[E]['req']>(
       endpoint: E,
       params: P,